+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/atomist-rugs/rug-upgrade/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/atomist-rugs/rug-upgrade/compare/0.2.0...HEAD
+
+## [0.2.0] - 2017-03-14
+
+[0.2.0]: https://github.com/atomist-rugs/rug-upgrade/compare/0.1.0...0.2.0
+
+Test release
+
+### Added
+
+-   TestsToTypeScript editor
+
+### Changed
+
+-   Updated to rug 0.13.0, including tests using TestsToTypeScript
 
 ## [0.1.0] - 2017-01-01